requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-analytics-data==0.18.8

# PostHog Analytics
posthog==3.1.0

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3

# Data processing and validation
pandas==2.1.3
numpy==1.25.2
isodate==0.6.1

# Background Jobs
apscheduler==3.10.4

# Rate limiting
slowapi==0.1.9

# Logging
loguru==0.7.2
